changed the code to check between which hours you can call the background service

+++ MM Systems/PanoramaApp3/PanoramaApp3/Pages/Settings.xaml.cs
@@ -141,11 +141,12 @@ namespace PanoramaApp3
             MessageBox.Show("Je bent nu uitgelogd");
             User.ID = 0;
             User.Settings["ID"] = User.ID;
-            try
+            if (ScheduledActionService.Find("ms")!=null)
+            if (ScheduledActionService.Find("ms")!=null)
             {
                 ScheduledActionService.Remove("ms");
             }
-            catch (InvalidOperationException exception)
+            else
             {
                 
             }
